finance-app/src/backend/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25

# 认证
python-jose[cryptography]==3.3.4
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 配置管理
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6
